% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{as.exdqlm}
\alias{as.exdqlm}
\title{\code{exdqlm} objects}
\usage{
as.exdqlm(m)
}
\arguments{
\item{m}{a list containing named elements m0, C0, FF and GG.}
}
\value{
A object of class "\code{exdqlm}" containing the state space model components:
\itemize{
\item FF - Observational vector.
\item GG - Evolution matrix.
\item m0 - Prior mean of the state vector.
\item C0 - Prior covariance of the state vector.
}
}
\description{
\code{as.exdqlm} attempts to turn a list into an \code{exdqlm} object. Works for time-invariant \code{dlm} objects created using the \pkg{dlm} package.
}
