% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTflow.R
\docType{data}
\name{BTflow}
\alias{BTflow}
\title{Monthly streamflow at the Big Trees gauge}
\format{
A monthly time series (\code{ts}) of length 471, spanning January 1987
through March 2026.
}
\source{
U.S. Geological Survey, National Water Information System
(USGS Water Data for the Nation), \url{https://waterdata.usgs.gov/}.
}
\usage{
BTflow
}
\description{
Observed monthly mean streamflow, in cubic feet per second, at the Big
Trees gauge on the San Lorenzo River in Santa Cruz County, California.
The monthly values were obtained by averaging the daily USGS streamflow
observations within each calendar month.
}
\references{
U.S. Geological Survey (2016). National Water Information
System data available on the World Wide Web (USGS Water Data for the
Nation). \url{https://waterdata.usgs.gov/}.
}
\keyword{datasets}
