% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTprec.R
\docType{data}
\name{BTprec}
\alias{BTprec}
\title{Monthly time-series of precipitation at Big Trees.}
\format{
A time series of length 471.
}
\source{
\url{https://waterdata.usgs.gov/}
}
\usage{
BTprec
}
\description{
Average monthly precipitation (mm) at the Big Trees on the San Lorenzo river in Santa Cruz County, CA from January 1987 through March 2026.
}
\references{
U.S. Geological Survey (2016). National Water Information System data available on the World Wide Web (USGS Water Data for the Nation). \url{https://waterdata.usgs.gov/}.
}
\keyword{datasets}
