% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal.r
\name{dexal}
\alias{dexal}
\title{Density Function for the Extended Asymmetric Laplace (exAL) Distribution}
\usage{
dexal(x, p0 = 0.5, mu = 0, sigma = 1, gamma = 0, log = FALSE)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{p0}{Probability level associated with the quantile parameterization. Must be in (0,1). Default is 0.5.}

\item{mu}{Location parameter. Default is 0.}

\item{sigma}{Scale parameter. Must be strictly positive. Default is 1.}

\item{gamma}{Skewness parameter controlling asymmetry. Default is 0.}

\item{log}{Logical; if TRUE, returns the log-density. Default is FALSE.}
}
\value{
A numeric vector of probability densities.
}
\description{
Computes the probability density function (PDF) for the Extended Asymmetric Laplace (exAL) distribution.
}
\details{
The exAL distribution extends the Asymmetric Laplace (AL) distribution by incorporating an additional skewness parameter \code{gamma}.
This distribution is useful for modeling heavy-tailed asymmetric data.
}
\examples{
dexal(0)
dexal(1, p0 = 0.75, mu = 0, sigma = 2, gamma = 1)
dexal(seq(-3, 3, by = 0.1), p0 = 0.3, mu = 0, sigma = 1, gamma = -0.5)

}
