% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{diagnostic}
\alias{diagnostic}
\title{Diagnostic Generic}
\usage{
diagnostic(object, ...)
}
\arguments{
\item{...}{Additional arguments passed to specific methods.}

\item{x}{An object of class \code{exdqlmFit}, \code{exdqlmForecast}, or
\code{exalStaticFit}.}
}
\value{
The output depends on the underlying method.
}
\description{
Calculates diagnostic metrics for a variety of objects.
}
