% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_require_fun.R
\name{.require_fun}
\alias{.require_fun}
\title{Internal helper: require a function to exist}
\usage{
.require_fun(fname, pkg = NULL)
}
\description{
Internal helper: require a function to exist
}
\keyword{internal}
