% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELIanoms.R
\docType{data}
\name{ELIanoms}
\alias{ELIanoms}
\title{Daily time-series of ELI anomalies.}
\format{
A time series of length 14965.
}
\source{
\url{https://portal.nersc.gov/archive/home/projects/cascade/www/ELI}
}
\usage{
ELIanoms
}
\description{
ELI anomalies on the daily time-scale from January 1, 1979 to December 31, 2019 with all February 29ths omitted.
}
\references{
Patricola, C.M., O’Brien, J.P., Risser, M.D. et al. \emph{Maximizing ENSO as a source of western US hydroclimate predictability.} Clim Dyn 54, 351–372 (2020). \doi{10.1007/s00382-019-05004-8}
}
\keyword{datasets}
