% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal_mcmc_fit.R
\name{exal_fit}
\alias{exal_fit}
\title{Fit exAL readout with either VB or MCMC}
\usage{
exal_fit(..., method = c("vb", "mcmc"), likelihood_family = c("exal", "al"))
}
\description{
Fit exAL readout with either VB or MCMC
}
