% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal_ldvb_engine.R
\name{exal_ldvb_engine}
\alias{exal_ldvb_engine}
\title{Internal LDVB engine (skeleton; returns exal_vb-compatible object)}
\usage{
exal_ldvb_engine(
  y,
  X,
  p0,
  gamma_bounds,
  vb_control,
  init,
  prior_gamma,
  prior_sigma,
  beta_prior_obj
)
}
\description{
Internal LDVB engine (skeleton; returns exal_vb-compatible object)
}
