% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal_ldvb_fit.R
\name{exal_ldvb_fit}
\alias{exal_ldvb_fit}
\title{Fit exAL readout with LDVB and pluggable beta prior}
\usage{
exal_ldvb_fit(
  y,
  X,
  p0,
  gamma_bounds,
  vb_control,
  init = list(),
  prior_gamma = list(mu0 = 0, s20 = 10),
  prior_sigma = list(a = 1, b = 1),
  beta_prior_obj = beta_prior("ridge", ridge = list(tau2 = 10000))
)
}
\description{
Fit exAL readout with LDVB and pluggable beta prior
}
