% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal_mcmc_fit.R
\name{exal_mcmc_posterior_draws}
\alias{exal_mcmc_posterior_draws}
\title{Draw posterior samples of (beta, sigma, gamma) from an exAL MCMC fit}
\usage{
exal_mcmc_posterior_draws(fit_exal, nd = NULL, seed = NULL)
}
\description{
Draw posterior samples of (beta, sigma, gamma) from an exAL MCMC fit
}
