% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal_mcmc_fit.R
\name{exal_mcmc_posterior_predict}
\alias{exal_mcmc_posterior_predict}
\title{Posterior predictive samples for an exAL MCMC fit}
\usage{
exal_mcmc_posterior_predict(
  fit_exal,
  X_new,
  nd = 1000L,
  chunk = 200L,
  draws = NULL,
  seed = NULL
)
}
\description{
Posterior predictive samples for an exAL MCMC fit
}
