% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdesn_vb.R
\name{exal_vb_posterior_draws}
\alias{exal_vb_posterior_draws}
\title{Draw posterior samples of (beta, sigma, gamma) from an exal_vb fit}
\usage{
exal_vb_posterior_draws(fit_exal, nd = 1000L)
}
\arguments{
\item{fit_exal}{Object returned by exal_static_LDVB()}

\item{nd}{Number of draws}
}
\value{
list(beta=nd x p, sigma=nd, gamma=nd)
}
\description{
Draw posterior samples of (beta, sigma, gamma) from an exal_vb fit
}
