% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmDiagnostics.R
\name{exdqlmDiagnostics}
\alias{exdqlmDiagnostics}
\title{exDQLM Diagnostics}
\usage{
exdqlmDiagnostics(
  m1,
  m2 = NULL,
  plot = TRUE,
  cols = c("red", "blue"),
  ref = NULL
)
}
\arguments{
\item{m1}{An object of class "\code{exdqlmLDVB}", "\code{exdqlmMCMC}", or
legacy "\code{exdqlmISVB}".}

\item{m2}{An optional additional object of class "\code{exdqlmLDVB}",
"\code{exdqlmMCMC}", or legacy "\code{exdqlmISVB}" to compare with \code{m1}.}

\item{plot}{Logical value indicating whether the following will be plotted for \code{m1} and \code{m2} (if provided): a qq-plot and ACF plot of the MAP one-step-ahead distribution sequence, and a time series plot of the standardized forecast errors. Default is \code{TRUE}.}

\item{cols}{Character vector of length 1 or 2 giving color(s) used to plot diagnostics. Default \code{c("red","blue")}.}

\item{ref}{Optional reference sample of size \code{length(m1$y)} from a standard
normal distribution. Used to compute the KL divergences.}
}
\value{
An object of class "\code{exdqlmDiagnostic}" containing the following:
\itemize{
\item \code{m1.uts} - The one-step-ahead distribution sequence of \code{m1}.
\item \code{m1.KL} - The KL divergence of \code{m1.uts} and a standard normal.
\item \code{m1.KL.flip} - The reversed ("flipped") KL divergence of \code{m1.uts}
and a standard normal.
\item \code{m1.CRPS} - The mean CRPS computed from posterior predictive draws.
\item \code{m1.pplc} - The posterior predictive loss criterion of \code{m1} based off the check loss function.
\item \code{m1.qq} - The ordered pairs of the qq-plot comparing \code{m1.uts} with a standard normal distribution.
\item \code{m1.acf} - The autocorrelations of \code{m1.uts} by lag.
\item \code{m1.rt} - Run-time of the original model \code{m1} in seconds.
\item \code{m1.msfe} - MAP standardized one-step-ahead forecast errors from the original model \code{m1}.
\item \code{y} - The original time-series used to fit \code{m1}.
}
If \code{m2} is provided, analogous results for \code{m2} are also included in the list.
}
\description{
The function computes the following for the model(s) provided: the posterior
predictive loss criterion based off the check loss, the CRPS from posterior
predictive draws, the one-step-ahead distribution sequence, and both the
forward and reversed KL divergences from normality. The function also plots
the following: the qq-plot and ACF plot corresponding to the one-step-ahead
distribution sequence, and a time series plot of the MAP standard forecast
errors.
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
old = options(exdqlm.max_iter = 15L)
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.95), dim.df = c(1),
                  gam.init = -3.5, sig.init = 15,
                  n.samp = 20, tol = 0.2, verbose = FALSE)
M0.diags = exdqlmDiagnostics(M0, plot = FALSE)
options(old)
}

}
