% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_beta.R
\name{\%||\%}
\alias{\%||\%}
\title{Construct beta prior object used by LDVB engine}
\usage{
x \%||\% y
}
\arguments{
\item{type}{"ridge" or "rhs"}

\item{ridge}{list(tau2=...)}

\item{rhs}{list(tau0, nu, s2, init_log_lambda, init_log_tau, init_log_c2,
rhs_maxit, rhs_reltol)}
}
\description{
Construct beta prior object used by LDVB engine
}
