% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exal_ldvb}
\alias{is.exal_ldvb}
\title{\code{exal_ldvb} objects}
\usage{
is.exal_ldvb(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exal_ldvb} tests if its argument is an \code{exal_ldvb} object.
}
