% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exal_mcmc}
\alias{is.exal_mcmc}
\title{\code{exal_mcmc} objects}
\usage{
is.exal_mcmc(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exal_mcmc} tests if its argument is an \code{exal_mcmc} object.
}
