% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalDiagnostics.R
\name{is.exalDiagnostic}
\alias{is.exalDiagnostic}
\title{\code{exalDiagnostic} objects}
\usage{
is.exalDiagnostic(x)
}
\arguments{
\item{x}{an \strong{R} object}
}
\description{
\code{is.exalDiagnostic} tests if its argument is an \code{exalDiagnostic}
object.
}
