% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticDiagnostic_methods.R
\name{is.exalStaticDiagnostic}
\alias{is.exalStaticDiagnostic}
\title{\code{exalStaticDiagnostic} objects}
\usage{
is.exalStaticDiagnostic(x)
}
\arguments{
\item{x}{an \strong{R} object}
}
\description{
\code{is.exalStaticDiagnostic} tests if its argument is an \code{exalStaticDiagnostic}
object.
}
