% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticFit_methods.R
\name{is.exalStaticFit}
\alias{is.exalStaticFit}
\title{\code{exalStaticFit} objects}
\usage{
is.exalStaticFit(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exalStaticFit} tests if its argument is a fitted static exAL
regression object, including MCMC and LDVB fits.
}
