% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticFit_generics.R
\name{is.exalStaticLDVB}
\alias{is.exalStaticLDVB}
\title{\code{exalStaticLDVB} objects}
\usage{
is.exalStaticLDVB(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exalStaticLDVB} tests if its argument is an \code{exalStaticLDVB} object.
}
