% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exalStaticMCMC}
\alias{is.exalStaticMCMC}
\title{\code{exalStaticMCMC} objects}
\usage{
is.exalStaticMCMC(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exalStaticMCMC} tests if its argument is an \code{exalStaticMCMC} object.
}
