% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exdqlm}
\alias{is.exdqlm}
\title{\code{exdqlm} objects}
\usage{
is.exdqlm(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exdqlm} tests if its argument is a \code{exdqlm} object.
}
