% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exdqlmDiagnostic}
\alias{is.exdqlmDiagnostic}
\title{\code{exdqlmDiagnostic} objects}
\usage{
is.exdqlmDiagnostic(x)
}
\arguments{
\item{x}{an \strong{R} object}
}
\description{
\code{is.exdqlmDiagnostic} tests if its argument is a \code{exdqlmDiagnostic} object.
}
