% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exdqlmFit}
\alias{is.exdqlmFit}
\title{\code{exdqlmFit} objects}
\usage{
is.exdqlmFit(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exdqlmFit} tests if its argument is a fitted dynamic
\code{exdqlm} object, including MCMC, LDVB, and legacy ISVB fits.
}
