% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmForecast_methods.R
\name{is.exdqlmForecast}
\alias{is.exdqlmForecast}
\title{\code{exdqlmForecast} objects}
\usage{
is.exdqlmForecast(x)
}
\arguments{
\item{x}{an \strong{R} object}
}
\description{
\code{is.exdqlmForecast} tests if its argument is a \code{exdqlmForecast} object.
}
