% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmForecastDiagnostic_methods.R
\name{is.exdqlmForecastDiagnostic}
\alias{is.exdqlmForecastDiagnostic}
\title{\code{exdqlmForecastDiagnostic} objects}
\usage{
is.exdqlmForecastDiagnostic(x)
}
\arguments{
\item{x}{an \strong{R} object.}
}
\description{
\code{is.exdqlmForecastDiagnostic} tests if its argument is an
\code{exdqlmForecastDiagnostic} object.
}
