% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_generics.R
\name{is.exdqlmISVB}
\alias{is.exdqlmISVB}
\title{\code{exdqlmISVB} objects}
\usage{
is.exdqlmISVB(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exdqlmISVB} tests if its argument is a \code{exdqlmISVB} object.
}
