% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_methods.R
\name{is.exdqlmLDVB}
\alias{is.exdqlmLDVB}
\title{\code{exdqlmLDVB} objects}
\usage{
is.exdqlmLDVB(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exdqlmLDVB} tests if its argument is a \code{exdqlmLDVB} object.
}
