% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmSynthesis_generics.R
\name{is.exdqlmSynthesis}
\alias{is.exdqlmSynthesis}
\title{\code{exdqlmSynthesis} objects}
\usage{
is.exdqlmSynthesis(x)
}
\arguments{
\item{x}{an \strong{R} object}
}
\description{
\code{is.exdqlmSynthesis} tests if its argument is an
\code{exdqlmSynthesis} object returned by \code{\link{quantileSynthesis}}.
}
