% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monELI.R
\docType{data}
\name{monELI}
\alias{monELI}
\title{Monthly time-series of ELI.}
\format{
A time series of length 936.
}
\source{
\url{https://portal.nersc.gov/archive/home/projects/cascade/www/ELI}
}
\usage{
monELI
}
\description{
ELI on the monthly time scale from 1937 through 2014.
}
\references{
Patricola, C.M., O’Brien, J.P., Risser, M.D. et al. \emph{Maximizing ENSO as a source of western US hydroclimate predictability.} Clim Dyn 54, 351–372 (2020). \doi{10.1007/s00382-019-05004-8}
}
\keyword{datasets}
