% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nino34.R
\docType{data}
\name{nino34}
\alias{nino34}
\title{Monthly Niño 3.4 Index.}
\format{
A time series of length 936.
}
\source{
\url{https://psl.noaa.gov/gcos_wgsp/Timeseries/Nino34/}
}
\usage{
nino34
}
\description{
Monthly Niño 3.4 sea surface temperature (SST) Index time series.
}
\references{
Rayner N. A., D. E. Parker, E. B. Horton, C. K. Folland, L. V. Alexander, D. P. Rowell, E. C. Kent, A. Kaplan. \emph{Global analyses of sea surface temperature, sea ice, and night marine air temperature since the late nineteenth century}. J. Geophys. Res., 108 (D14), 4407 (2003). \doi{10.1029/2002JD002670}
}
\keyword{datasets}
