% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal.r
\name{pexal}
\alias{pexal}
\title{Cumulative Distribution Function (CDF) for the exAL Distribution}
\usage{
pexal(
  q,
  p0 = 0.5,
  mu = 0,
  sigma = 1,
  gamma = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{Numeric vector of quantiles.}

\item{p0}{Probability level associated with the quantile parameterization. Must be in (0,1). Default is 0.5.}

\item{mu}{Location parameter. Default is 0.}

\item{sigma}{Scale parameter. Must be strictly positive. Default is 1.}

\item{gamma}{Skewness parameter controlling asymmetry. Default is 0.}

\item{lower.tail}{Logical; if TRUE (default), returns P(X ≤ q), otherwise P(X > q).}

\item{log.p}{Logical; if TRUE, returns log-probabilities. Default is FALSE.}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Computes the cumulative probability for the Extended Asymmetric Laplace (exAL) distribution.
}
\examples{
pexal(0)
pexal(1, p0 = 0.75, mu = 0, sigma = 2, gamma = 1)
pexal(seq(-3, 3, by = 0.1), p0 = 0.3, mu = 0, sigma = 1, gamma = -0.5)

}
