% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalDiagnostics.R
\name{plot.exalDiagnostic}
\alias{plot.exalDiagnostic}
\title{Plot Method for \code{exalDiagnostic} Objects}
\usage{
\method{plot}{exalDiagnostic}(x, cols = c("red", "blue"), ...)
}
\arguments{
\item{x}{An \code{exalDiagnostic} object.}

\item{cols}{Character vector of length 1 or 2 giving color(s) used to plot
diagnostics.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plot Method for \code{exalDiagnostic} Objects
}
