% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmDiagnostic_methods.R
\name{plot.exdqlmDiagnostic}
\alias{plot.exdqlmDiagnostic}
\title{Plot Method for \code{exdqlmDiagnostic} Objects}
\usage{
\method{plot}{exdqlmDiagnostic}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmDiagnostic} object.}

\item{...}{Additional graphical arguments. The optional \code{cols} element
controls the colors used for the first and second model when comparing two
fits.}
}
\value{
Invisibly returns \code{x}.
}
\description{
This function produces the QQ plot and
ACF plot corresponding to the one-step-ahead distribution sequence, together
with a time series plot of the MAP standard forecast errors.
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
old = options(exdqlm.max_iter = 15L)
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.95), dim.df = c(1),
                  gam.init = -3.5, sig.init = 15,
                  n.samp = 20, tol = 0.2, verbose = FALSE)
M0.diags = exdqlmDiagnostics(M0)
plot(M0.diags)
options(old)
}

}
