% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_generics.R
\name{plot.exdqlmMCMC}
\alias{plot.exdqlmMCMC}
\title{Plot Method for \code{exdqlmMCMC} Objects}
\usage{
\method{plot}{exdqlmMCMC}(x, type = c("quantile", "component", "state"), index = NULL, ...)
}
\arguments{
\item{x}{An \code{exdqlmMCMC} object.}

\item{type}{Character string specifying \code{"quantile"}, \code{"component"},
or \code{"state"}.}

\item{index}{Required for \code{type = "component"} or \code{type = "state"}.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for \code{exdqlmMCMC} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M2 = exdqlmMCMC(y, p0=0.85, model, df = c(0.98), dim.df = c(1),
                gam.init = -3.5, sig.init = 15,
                n.burn = 20, n.mcmc = 20,
                init.from.vb = FALSE, verbose = FALSE)
plot(M2)                
}

}
