% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdesn_vb.R
\name{predict_mu.qdesn_fit}
\alias{predict_mu.qdesn_fit}
\title{Predict in-sample p0-quantiles (mu_t) for a fitted Q-DESN}
\usage{
predict_mu.qdesn_fit(object)
}
\arguments{
\item{object}{A \code{qdesn_fit} object.}
}
\value{
A numeric vector \eqn{\hat\mu_t} aligned with \code{object$y_fit}.
}
\description{
Predict in-sample p0-quantiles (mu_t) for a fitted Q-DESN
}
