% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exal_ldvb}
\alias{print.exal_ldvb}
\title{Print Method for \code{exal_ldvb} Objects}
\usage{
\method{print}{exal_ldvb}(x, ...)
}
\arguments{
\item{x}{An \code{exal_ldvb} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exal_ldvb} Objects
}
