% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exal_mcmc}
\alias{print.exal_mcmc}
\title{Print Method for \code{exal_mcmc} Objects}
\usage{
\method{print}{exal_mcmc}(x, ...)
}
\arguments{
\item{x}{An \code{exal_mcmc} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exal_mcmc} Objects
}
