% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalDiagnostics.R
\name{print.exalDiagnostic}
\alias{print.exalDiagnostic}
\title{Print Method for \code{exalDiagnostic} Objects}
\usage{
\method{print}{exalDiagnostic}(x, ...)
}
\arguments{
\item{x}{An \code{exalDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exalDiagnostic} Objects
}
