% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticFit_methods.R
\name{print.exalStaticFit}
\alias{print.exalStaticFit}
\title{Print Method for \code{exalStaticFit} Objects}
\usage{
\method{print}{exalStaticFit}(x, ...)
}
\arguments{
\item{x}{A fitted static \code{exalStaticFit} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exalStaticFit} Objects
}
