% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exdqlm}
\alias{print.exdqlm}
\title{Print exDQLM model details}
\usage{
\method{print}{exdqlm}(x, ...)
}
\arguments{
\item{x}{a \code{exdqlm} object.}

\item{...}{further arguments (unused).}
}
\description{
Print the details of the exDQLM model.
}
