% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmDiagnostic_generics.R
\name{print.exdqlmDiagnostic}
\alias{print.exdqlmDiagnostic}
\title{Print Method for \code{exdqlmDiagnostic} Objects}
\usage{
\method{print}{exdqlmDiagnostic}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmDiagnostic} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
old = options(exdqlm.max_iter = 15L)
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.95), dim.df = c(1),
                  gam.init = -3.5, sig.init = 15,
                  n.samp = 20, tol = 0.2, verbose = FALSE)
M0.diags = exdqlmDiagnostics(M0)
print(M0.diags)
options(old)
}

}
