% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exdqlmMCMC}
\alias{print.exdqlmMCMC}
\title{Print Method for \code{exdqlmMCMC} Objects}
\usage{
\method{print}{exdqlmMCMC}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmMCMC} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmMCMC} Objects
}
