% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmSynthesis_methods.R
\name{print.exdqlmSynthesis}
\alias{print.exdqlmSynthesis}
\title{Print Method for \code{exdqlmSynthesis} Objects}
\usage{
\method{print}{exdqlmSynthesis}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmSynthesis} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmSynthesis} Objects
}
