% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdesn_design_only.R
\name{qdesn_build_design}
\alias{qdesn_build_design}
\title{Build Q-DESN design matrix (reservoir -> readout features) without fitting readout}
\usage{
qdesn_build_design(y, desn_args, drop = NULL)
}
\description{
Build Q-DESN design matrix (reservoir -> readout features) without fitting readout
}
