% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exal.r
\name{qexal}
\alias{qexal}
\title{Quantile Function for the exAL Distribution}
\usage{
qexal(
  p,
  p0 = 0.5,
  mu = 0,
  sigma = 1,
  gamma = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{p}{Numeric vector of probabilities. Must be in (0,1).}

\item{p0}{Probability level associated with the quantile parameterization. Must be in (0,1). Default is 0.5.}

\item{mu}{Location parameter. Default is 0.}

\item{sigma}{Scale parameter. Must be strictly positive. Default is 1.}

\item{gamma}{Skewness parameter controlling asymmetry. Default is 0.}

\item{lower.tail}{Logical; if TRUE (default), returns the lower-tail quantile. If FALSE, returns the upper-tail quantile.}

\item{log.p}{Logical; if TRUE, probabilities are given in log-scale. Default is FALSE.}
}
\value{
A numeric vector of quantiles.
}
\description{
Computes the quantile function (inverse CDF) for the Extended Asymmetric Laplace (exAL) distribution.
}
\examples{
qexal(0.5)
qexal(0.95, p0 = 0.75, mu = 0, sigma = 2, gamma = 1)
qexal(seq(0.1, 0.9, by = 0.1), p0 = 0.3, mu = 0, sigma = 1, gamma = -0.5)

}
