% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIVTmag.R
\docType{data}
\name{scIVTmag}
\alias{scIVTmag}
\title{Time series of daily average magnitude IVT in Santa Cruz, CA.}
\format{
A time series of length 14965.
}
\source{
\url{https://cds.climate.copernicus.eu}
}
\usage{
scIVTmag
}
\description{
ECMWF Re-Analysis 5 (ERA5) daily average magnitude IVT in Santa Cruz, CA (approximately 22 N, 122 W)
from January 1, 1979 to December 31, 2019 with all February 29ths omitted.
}
\references{
Hersbach, H, Bell, B, Berrisford, P, et al. \emph{The ERA5 global reanalysis.} Q J R Meteorol Soc. 2020; 146: 1999– 2049. \doi{10.1002/qj.3803}
}
\keyword{datasets}
