% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exal_ldvb}
\alias{summary.exal_ldvb}
\title{Summary Method for \code{exal_ldvb} Objects}
\usage{
\method{summary}{exal_ldvb}(object, ...)
}
\arguments{
\item{object}{An \code{exal_ldvb} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exal_ldvb} Objects
}
