% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exal_mcmc}
\alias{summary.exal_mcmc}
\title{Summary Method for \code{exal_mcmc} Objects}
\usage{
\method{summary}{exal_mcmc}(object, ...)
}
\arguments{
\item{object}{An \code{exal_mcmc} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exal_mcmc} Objects
}
