% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalDiagnostics.R
\name{summary.exalDiagnostic}
\alias{summary.exalDiagnostic}
\title{Summary Method for \code{exalDiagnostic} Objects}
\usage{
\method{summary}{exalDiagnostic}(object, ...)
}
\arguments{
\item{object}{An \code{exalDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exalDiagnostic} Objects
}
