% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticDiagnostics.R
\name{summary.exalStaticDiagnostic}
\alias{summary.exalStaticDiagnostic}
\alias{summary.exalDiagnostic}
\title{Summary Method for \code{exalStaticDiagnostic} Objects}
\usage{
\method{summary}{exalStaticDiagnostic}(object, ...)

\method{summary}{exalDiagnostic}(object, ...)
}
\arguments{
\item{object}{An \code{exalStaticDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exalStaticDiagnostic} Objects
}
