% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exalStaticFit}
\alias{summary.exalStaticFit}
\title{Summary Method for \code{exalStaticFit} Objects}
\usage{
\method{summary}{exalStaticFit}(object, max.coef = 6L, ...)
}
\arguments{
\item{object}{A fitted static \code{exalStaticFit} object.}

\item{max.coef}{Maximum number of coefficients to print in the coefficient
summary table.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exalStaticFit} Objects
}
