% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exalStaticMCMC}
\alias{summary.exalStaticMCMC}
\title{Summary Method for \code{exalStaticMCMC} Objects}
\usage{
\method{summary}{exalStaticMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{exalStaticMCMC} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exalStaticMCMC} Objects
}
