% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlm}
\alias{summary.exdqlm}
\title{Summary exDQLM model details}
\usage{
\method{summary}{exdqlm}(object, ...)
}
\arguments{
\item{object}{a \code{exdqlm} object.}

\item{...}{further arguments (unused).}
}
\description{
Print the details of the exDQLM model.
}
