% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlmFit}
\alias{summary.exdqlmFit}
\title{Summary Method for \code{exdqlmFit} Objects}
\usage{
\method{summary}{exdqlmFit}(object, ...)
}
\arguments{
\item{object}{A fitted dynamic \code{exdqlmFit} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmFit} Objects
}
