% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmForecast_methods.R
\name{summary.exdqlmForecast}
\alias{summary.exdqlmForecast}
\title{Summary Method for \code{exdqlmForecast} Objects}
\usage{
\method{summary}{exdqlmForecast}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmForecast} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmForecast} Objects
}
\examples{
\donttest{
 data("scIVTmag", package = "exdqlm")
 old = options(exdqlm.max_iter = 15L)
 y = scIVTmag[1:60]
 model = polytrendMod(1, stats::quantile(y, 0.85), 10)
 M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                  dqlm.ind = TRUE, sig.init = 15,
                  n.samp = 20, tol = 0.2, verbose = FALSE)
 M0.forecast = predict(M0, start.t = 50, k = 5)
 summary(M0.forecast)
 options(old)
}

}
