% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmForecastDiagnostics.R
\name{summary.exdqlmForecastDiagnostic}
\alias{summary.exdqlmForecastDiagnostic}
\title{Summary Method for \code{exdqlmForecastDiagnostic} Objects}
\usage{
\method{summary}{exdqlmForecastDiagnostic}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmForecastDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmForecastDiagnostic} Objects
}
