% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_methods.R
\name{summary.exdqlmLDVB}
\alias{summary.exdqlmLDVB}
\title{Summary Method for \code{exdqlmLDVB} Objects}
\usage{
\method{summary}{exdqlmLDVB}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmLDVB} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmLDVB} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
old = options(exdqlm.max_iter = 15L)
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                   dqlm.ind = TRUE, sig.init = 15,
                   n.samp = 20, tol = 0.2, verbose = FALSE)
summary(M0)
options(old)
}

}
