% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlmMCMC}
\alias{summary.exdqlmMCMC}
\title{Summary Method for \code{exdqlmMCMC} Objects}
\usage{
\method{summary}{exdqlmMCMC}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmMCMC} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmMCMC} Objects
}
\examples{
\donttest{
y = scIVTmag[1:100]
model = polytrendMod(1,quantile(y,0.85),10)
M2 = exdqlmMCMC(y,p0=0.85,model,df=c(0.98),dim.df = c(1),
                gam.init=-3.5,sig.init=15,
                n.burn=100,n.mcmc=150)
summary(M2)                
}

}
