% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTflow.R
\docType{data}
\name{BTflow}
\alias{BTflow}
\title{Monthly time-series of water flow at Big Tree water gauge.}
\format{
A time series of length 936.
}
\source{
\url{https://waterdata.usgs.gov/nwis/}
}
\usage{
BTflow
}
\description{
Average monthly natural water flow (cubic feet per second) at the Big Tree gauge of the San Lorenzo river in Santa Cruz County, CA from 1937 through 2014.
}
\references{
U.S. Geological Survey (2016). National Water Information System data available on the World Wide Web (USGS Water Data for the Nation). \url{https://waterdata.usgs.gov/nwis/}.
}
\keyword{datasets}
