% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlm-package.R
\docType{package}
\name{exdqlm-package}
\alias{exdqlm}
\alias{exdqlm-package}
\title{exdqlm: Extended Dynamic Quantile Linear Models}
\description{
Routines for Bayesian estimation and analysis of dynamic quantile linear models utilizing the extended asymmetric Laplace error distribution, also known as extended dynamic quantile linear models (exDQLM) described in Barata et al (2020) \doi{10.1214/21-AOAS1497}.
}
\author{
\strong{Maintainer}: Raquel Barata \email{raquel.a.barata@gmail.com}

Authors:
\itemize{
  \item Antonio Aguirre
}

Other contributors:
\itemize{
  \item Raquel Prado [thesis advisor]
  \item Bruno Sanso [thesis advisor]
}

}
\keyword{internal}
