% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlm-package.R
\docType{package}
\name{exdqlm-package}
\alias{exdqlm}
\alias{exdqlm-package}
\title{exdqlm: Extended Dynamic Quantile Linear Models}
\description{
Routines for Bayesian estimation and analysis of dynamic quantile linear models
using the extended asymmetric Laplace error distribution (exDQLM).
}
\section{Runtime options}{

\itemize{
\item \code{options(exdqlm.use_cpp_kf = TRUE|FALSE)} – C++ Kalman bridge (optional; default TRUE).
\item \code{options(exdqlm.use_cpp_samplers = TRUE|FALSE)} – C++ samplers (optional; default FALSE).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AntonioAPDL/exdqlm}
  \item Report bugs at \url{https://github.com/AntonioAPDL/exdqlm/issues}
}

}
\author{
\strong{Maintainer}: Raquel Barata \email{raquel.a.barata@gmail.com}

Authors:
\itemize{
  \item Antonio Aguirre
}

Other contributors:
\itemize{
  \item Raquel Prado [thesis advisor]
  \item Bruno Sanso [thesis advisor]
}

}
\keyword{package}
