% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmPlot.R
\name{exdqlmPlot}
\alias{exdqlmPlot}
\title{Plot exDQLM}
\usage{
exdqlmPlot(m1, add = FALSE, col = "purple", cr.percent = 0.95)
}
\arguments{
\item{m1}{An object of class "\code{exdqlmLDVB}", "\code{exdqlmMCMC}", or
legacy "\code{exdqlmISVB}".}

\item{add}{Logical value indicating whether the dynamic quantile will be added to existing plot. Default is \code{FALSE}.}

\item{col}{Character vector of length 1 giving color of the dynamic quantile to be plotted. Default is \code{purple}.}

\item{cr.percent}{Numeric in \code{(0, 1)} indicating the probability mass for the credible
intervals (e.g., \code{0.95}). Default \code{0.95}.}
}
\value{
A list of the following is returned:
\itemize{
\item \code{map.quant} - MAP estimate of the dynamic quantile.
\item \code{lb.quant} - Lower bound of the 95\% CrIs of the dynamic quantile.
\item \code{ub.quant} - Upper bound of the 95\% CrIs of the dynamic quantile.
}
}
\description{
The function plots the MAP estimates and 95\% credible intervals (CrIs) of the dynamic quantile of an exDQLM.
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
y = scIVTmag[1:100]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                   gam.init = -3.5, sig.init = 15)
exdqlmPlot(M0, col = "blue")
}

}
