% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{is.exdqlmMCMC}
\alias{is.exdqlmMCMC}
\title{\code{exdqlmMCMC} objects}
\usage{
is.exdqlmMCMC(m)
}
\arguments{
\item{m}{an \strong{R} object}
}
\description{
\code{is.exdqlmMCMC} tests if its argument is a \code{exdqlmMCMC} object.
}
