% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticDiagnostics.R
\name{plot.exalStaticDiagnostic}
\alias{plot.exalStaticDiagnostic}
\title{Plot Method for \code{exalStaticDiagnostic} Objects}
\usage{
\method{plot}{exalStaticDiagnostic}(x, cols = c("red", "blue"), ...)
}
\arguments{
\item{x}{An \code{exalStaticDiagnostic} object.}

\item{cols}{Character vector of length 1 or 2 giving color(s) used to plot
diagnostics.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plot Method for \code{exalStaticDiagnostic} Objects
}
