% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{plot.exdqlmForecast}
\alias{plot.exdqlmForecast}
\title{Plot Method for \code{exdqlmForecast} Objects}
\usage{
\method{plot}{exdqlmForecast}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmForecast} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Plot Method for \code{exdqlmForecast} Objects
}
\examples{
\donttest{
 data("scIVTmag", package = "exdqlm")
 old = options(exdqlm.max_iter = 15L)
 y = scIVTmag[1:60]
 model = polytrendMod(1, stats::quantile(y, 0.85), 10)
 M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                  gam.init = -3.5, sig.init = 15,
                  n.samp = 20, tol = 0.2, verbose = FALSE)
 M0.forecast = exdqlmForecast(start.t = 50, k = 5, m1 = M0)
 plot(M0.forecast)
 options(old)
}

}
