% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_generics.R
\name{plot.exdqlmLDVB}
\alias{plot.exdqlmLDVB}
\title{Plot Method for \code{exdqlmLDVB} Objects}
\usage{
\method{plot}{exdqlmLDVB}(x, type = c("quantile", "component", "state"), index = NULL, ...)
}
\arguments{
\item{x}{An \code{exdqlmLDVB} object.}

\item{type}{Character string specifying \code{"quantile"}, \code{"component"},
or \code{"state"}.}

\item{index}{Required for \code{type = "component"} or \code{type = "state"}.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for \code{exdqlmLDVB} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
old = options(exdqlm.max_iter = 15L)
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                   gam.init = -3.5, sig.init = 15,
                   n.samp = 20, tol = 0.2, verbose = FALSE)
plot(M0)
options(old)
}

}
