% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exalStaticFit_generics.R
\name{print.exalStaticMCMC}
\alias{print.exalStaticMCMC}
\title{Print Method for \code{exalStaticMCMC} Objects}
\usage{
\method{print}{exalStaticMCMC}(x, ...)
}
\arguments{
\item{x}{An \code{exalStaticMCMC} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exalStaticMCMC} Objects
}
