% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exdqlmDiagnostic}
\alias{print.exdqlmDiagnostic}
\title{Print Method for \code{exdqlmDiagnostic} Objects}
\usage{
\method{print}{exdqlmDiagnostic}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmDiagnostic} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmDiagnostic} Objects
}
\examples{
\donttest{
y = scIVTmag[1:100]
model = polytrendMod(1,mean(y),10)
M0 = exdqlmISVB(y,p0=0.85,model,df=c(0.95),dim.df = c(1),
                  gam.init=-3.5,sig.init=15)
M0.diags = exdqlmDiagnostics(M0,plot=FALSE)
print(M0.diags)
}

}
