% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{print.exdqlmISVB}
\alias{print.exdqlmISVB}
\title{Print Method for \code{exdqlmISVB} Objects}
\usage{
\method{print}{exdqlmISVB}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmISVB} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmISVB} Objects
}
