% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqlmFit_generics.R
\name{print.exdqlmMCMC}
\alias{print.exdqlmMCMC}
\title{Print Method for \code{exdqlmMCMC} Objects}
\usage{
\method{print}{exdqlmMCMC}(x, ...)
}
\arguments{
\item{x}{An \code{exdqlmMCMC} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for \code{exdqlmMCMC} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
y = scIVTmag[1:60]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M2 = exdqlmMCMC(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                gam.init = -3.5, sig.init = 15,
                n.burn = 20, n.mcmc = 20,
                init.from.vb = FALSE, verbose = FALSE)
print(M2)                
}

}
