% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exalStaticLDVB}
\alias{summary.exalStaticLDVB}
\title{Summary Method for \code{exalStaticLDVB} Objects}
\usage{
\method{summary}{exalStaticLDVB}(object, ...)
}
\arguments{
\item{object}{An \code{exalStaticLDVB} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exalStaticLDVB} Objects
}
