% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlmForecast}
\alias{summary.exdqlmForecast}
\title{Summary Method for \code{exdqlmForecast} Objects}
\usage{
\method{summary}{exdqlmForecast}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmForecast} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmForecast} Objects
}
\examples{
\donttest{
 y <- scIVTmag[1:100]
 model <- polytrendMod(1, stats::quantile(y, 0.85), 10)
 M0 <- exdqlmISVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                  gam.init = -3.5, sig.init = 15)
 M0.forecast = exdqlmForecast(start.t = 90, k = 10, m1 = M0)
 summary(M0.forecast)
}

}
