% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlmISVB}
\alias{summary.exdqlmISVB}
\title{Summary Method for \code{exdqlmISVB} Objects}
\usage{
\method{summary}{exdqlmISVB}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmISVB} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmISVB} Objects
}
\examples{
\donttest{
y = scIVTmag[1:100]
model = polytrendMod(1,quantile(y,0.85),10)
M0 = exdqlmISVB(y,p0=0.85,model,df=c(0.98),dim.df = c(1),
                   gam.init=-3.5,sig.init=15)
summary(M0)
}

}
