% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_etc.R
\name{summary.exdqlmLDVB}
\alias{summary.exdqlmLDVB}
\title{Summary Method for \code{exdqlmLDVB} Objects}
\usage{
\method{summary}{exdqlmLDVB}(object, ...)
}
\arguments{
\item{object}{An \code{exdqlmLDVB} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for \code{exdqlmLDVB} Objects
}
\examples{
\donttest{
data("scIVTmag", package = "exdqlm")
y = scIVTmag[1:100]
model = polytrendMod(1, stats::quantile(y, 0.85), 10)
M0 = exdqlmLDVB(y, p0 = 0.85, model, df = c(0.98), dim.df = c(1),
                   gam.init = -3.5, sig.init = 15)
summary(M0)
}

}
